+++ src/classes/babies/SquirrelBaby.ts
@@ -4,13 +4,14 @@ import { g } from "../../globals";
 import { mod } from "../../mod";
 import { Baby } from "../Baby";
 
+/** This cannot be in a `num` property since it would grant that many walnuts. */
+const NUM_COLLECTIBLES = 5;
+
 /** Starts with Walnut (improved). */
 export class SquirrelBaby extends Baby {
   @CallbackCustom(ModCallbackCustom.POST_TRINKET_BREAK, TrinketType.WALNUT)
   postTrinketBreakWalnut(player: EntityPlayer): void {
-    const num = this.getAttribute("num");
-
-    repeat(num, () => {
+    repeat(NUM_COLLECTIBLES, () => {
       const position = g.r.FindFreePickupSpawnPosition(
         player.Position,
         1,